Moved getValue and setValue from PersistentResource to EntityDictionary
