+++ elide-core/src/main/java/com/yahoo/elide/core/PersistentResource.java
@@ -41,17 +41,13 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.Sets;
 
 import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
 
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.Serializable;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.AccessibleObject;
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -340,7 +336,7 @@ public class PersistentResource<T> implements com.yahoo.elide.security.Persisten
      */
     public boolean updateAttribute(String fieldName, Object newVal) {
         Class<?> fieldClass = dictionary.getType(getResourceClass(), fieldName);
-        newVal =  coerce(newVal, fieldName, fieldClass);
+        newVal =  dictionary.coerce(obj, newVal, fieldName, fieldClass);
         Object val = getValueUnchecked(fieldName);
         checkFieldAwareDeferPermissions(UpdatePermission.class, fieldName, newVal, val);
         if (!Objects.equals(val, newVal)) {
@@ -1423,112 +1419,11 @@ public class PersistentResource<T> implements com.yahoo.elide.security.Persisten
      * @param value the value to set
      */
     protected void setValue(String fieldName, Object value) {
-        Class<?> targetClass = obj.getClass();
         final Object original = getValueUnchecked(fieldName);
-        try {
-            Class<?> fieldClass = dictionary.getType(targetClass, fieldName);
-            String realName = dictionary.getNameFromAlias(obj, fieldName);
-            fieldName = (realName != null) ? realName : fieldName;
-            String setMethod = "set" + StringUtils.capitalize(fieldName);
-            Method method = EntityDictionary.findMethod(targetClass, setMethod, fieldClass);
-            method.invoke(obj, coerce(value, fieldName, fieldClass));
-        } catch (IllegalAccessException e) {
-            throw new InvalidAttributeException(fieldName, type, e);
-        } catch (InvocationTargetException e) {
-            throw handleInvocationTargetException(e);
-        } catch (IllegalArgumentException | NoSuchMethodException noMethod) {
-            AccessibleObject accessor = dictionary.getAccessibleObject(obj, fieldName);
-            if (accessor != null && accessor instanceof Field) {
-                Field field = (Field) accessor;
-                try {
-                    field.set(obj, coerce(value, fieldName, field.getType()));
-                } catch (IllegalAccessException noField) {
-                    throw new InvalidAttributeException(fieldName, type, noField);
-                }
-            } else {
-                throw new InvalidAttributeException(fieldName, type);
-            }
-        }
+
+        dictionary.setValue(obj, fieldName, value);
 
         triggerUpdate(fieldName, original, value);
-    }
-
-    /**
-     * Coerce provided value into expected class type.
-     *
-     * @param value provided value
-     * @param fieldName the field name
-     * @param fieldClass expected class type
-     * @return coerced value
-     */
-    private Object coerce(Object value, String fieldName, Class<?> fieldClass) {
-        if (fieldClass != null && Collection.class.isAssignableFrom(fieldClass) && value instanceof Collection) {
-            return coerceCollection((Collection) value, fieldName, fieldClass);
-        }
-
-        if (fieldClass != null && Map.class.isAssignableFrom(fieldClass) && value instanceof Map) {
-            return coerceMap((Map<?, ?>) value, fieldName, fieldClass);
-        }
-
-        return CoerceUtil.coerce(value, fieldClass);
-    }
-
-    private Collection coerceCollection(Collection<?> values, String fieldName, Class<?> fieldClass) {
-        Class<?> providedType = dictionary.getParameterizedType(obj, fieldName);
-
-        // check if collection is of and contains the correct types
-        if (fieldClass.isAssignableFrom(values.getClass())) {
-            boolean valid = true;
-            for (Object member : values) {
-                if (member != null && !providedType.isAssignableFrom(member.getClass())) {
-                    valid = false;
-                    break;
-                }
-            }
-            if (valid) {
-                return values;
-            }
-        }
-
-        ArrayList<Object> list = new ArrayList<>(values.size());
-        for (Object member : values) {
-            list.add(CoerceUtil.coerce(member, providedType));
-        }
-
-        if (Set.class.isAssignableFrom(fieldClass)) {
-            return new LinkedHashSet<>(list);
-        }
-
-        return list;
-    }
-
-    private Map coerceMap(Map<?, ?> values, String fieldName, Class<?> fieldClass) {
-        Class<?> keyType = dictionary.getParameterizedType(obj, fieldName, 0);
-        Class<?> valueType = dictionary.getParameterizedType(obj, fieldName, 1);
-
-        // Verify the existing Map
-        if (isValidParameterizedMap(values, keyType, valueType)) {
-            return values;
-        }
-
-        LinkedHashMap<Object, Object> result = new LinkedHashMap<>(values.size());
-        for (Map.Entry<?, ?> entry : values.entrySet()) {
-            result.put(CoerceUtil.coerce(entry.getKey(), keyType), CoerceUtil.coerce(entry.getValue(), valueType));
-        }
-
-        return result;
-    }
-
-    private boolean isValidParameterizedMap(Map<?, ?> values, Class<?> keyType, Class<?> valueType) {
-        for (Map.Entry<?, ?> entry : values.entrySet()) {
-            Object key = entry.getKey();
-            Object value = entry.getValue();
-            if ((key != null && !keyType.isAssignableFrom(key.getClass()))
-                    || (value != null && !valueType.isAssignableFrom(value.getClass()))) {
-                return false;
-            }
-        }
-        return true;
     }
 
     /**
@@ -1540,24 +1435,8 @@ public class PersistentResource<T> implements com.yahoo.elide.security.Persisten
      */
     public static Object getValue(Object target, String fieldName, RequestScope requestScope) {
         EntityDictionary dictionary = requestScope.getDictionary();
-        AccessibleObject accessor = dictionary.getAccessibleObject(target, fieldName);
-        try {
-            if (accessor instanceof Method) {
-                // Pass RequestScope into @Computed fields if requested
-                if (dictionary.isMethodRequestScopeable(target, (Method) accessor)) {
-                    return ((Method) accessor).invoke(target, requestScope);
-                }
-                return ((Method) accessor).invoke(target);
-            }
-            if (accessor instanceof Field) {
-                return ((Field) accessor).get(target);
-            }
-        } catch (IllegalAccessException e) {
-            throw new InvalidAttributeException(fieldName, dictionary.getJsonAliasFor(target.getClass()), e);
-        } catch (InvocationTargetException e) {
-            throw handleInvocationTargetException(e);
-        }
-        throw new InvalidAttributeException(fieldName, dictionary.getJsonAliasFor(target.getClass()));
+
+        return dictionary.getValue(target, fieldName, requestScope);
     }
 
     /**
